Make the zpool.cache copy optional. Its possible a zpool has a non standard cache file location, or doesn't have one at all.
Fixes: #14

+++ backend/functions-upgrade.sh
@@ -82,7 +82,9 @@ unmount_upgrade()
   fi
 
   # Copy zpool.cache
-  rc_halt "cp /boot/zfs/zpool.cache ${FSMNT}/boot/zfs/"
+  if [ -e "/boot/zfs/zpool.cache" ] ; then
+   rc_halt "cp /boot/zfs/zpool.cache ${FSMNT}/boot/zfs/"
+  fi
 
   # Unmount FS
   umount_all_dir "${FSMNT}"
